+++ pkg/wikidata/merger.go
@@ -68,12 +68,6 @@ func MergePOIs(candidates []*model.POI, cfg *config.CategoriesConfig, logger *sl
 				// Since we sorted by quality, 'acc' is better or equal to 'cand'.
 				// We drop 'cand'.
 				isDuplicate = true
-				if logger != nil {
-					// slog.Debug("Merged POI",
-					// 	"kept", acc.DisplayName(), "kept_qid", acc.WikidataID, "kept_len", acc.WPArticleLength,
-					// 	"dropped", cand.DisplayName(), "dropped_qid", cand.WikidataID, "dropped_len", cand.WPArticleLength,
-					// 	"dist_m", distMeters, "threshold", mergeDist)
-				}
 				break
 			}
 		}